requirements.txt
crewai==0.130.0
fastapi==0.110.3
uvicorn==0.30.1
python-dotenv==1.0.1
google-generativeai==0.5.4
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
httpx==0.27.0
Jinja2==3.1.4
jsonschema==4.22.0
oauthlib==3.2.2